requirements.txt
pytest~=7.0.1
selenium~=4.1.3
pytest-xdist~=2.5.0
allure-pytest~=2.9.45